Update and rename fibonacci.java to Fibonacci.java
